+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `db.Topic().DescribeTopicConsumer()` method for displaying consumer information
+
 ## v3.84.1
 * Added session info into `trace.TableSessionBulkUpsertStartInfo`